plugin-template-0.1/plugin-shotgun/plugin-legacy/requirements.txt
beautifulsoup4 == 4.8.1
flask
dataiq-plugin == 0.11.1
datetime
pyyaml
pytest
pytest-cov
requests
